+++ benches/signature_var_generator.rs
@@ -64,7 +64,8 @@ impl SigVarGenCircuit {
 
 impl Circuit for SigVarGenCircuit {
     fn circuit(&self, composer: &mut Composer) -> Result<(), PlonkError> {
-        let (u, r) = self.signature.append(composer);
+        let u = composer.append_witness(*self.signature.u());
+        let r = composer.append_point(self.signature.R());
 
         let pk = composer.append_point(self.pk.public_key());
         let generator = composer.append_point(self.pk.generator());